requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
tensorflow==2.15.0
scikit-learn==1.3.2
joblib==1.3.2
supabase==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6